Add helpers.js (front-end) to refactor the code
